// src/app/screens/productsPage/ChosenProduct.tsx
import React, { useEffect } from "react";
import { Container, Stack, Box } from "@mui/material";
import { Swiper, SwiperSlide } from "swiper/react";
import RemoveRedEyeIcon from "@mui/icons-material/RemoveRedEye";
import Divider from "../../components/divider";
import Button from "@mui/material/Button";
import Rating from "@mui/material/Rating";
import { useParams } from "react-router-dom"; // @ts-ignore
import { FreeMode, Navigation, Thumbs } from "swiper";
import "swiper/css";
import "swiper/css/free-mode";
import "swiper/css/navigation";
import "swiper/css/thumbs";

import {  Dispatch } from "@reduxjs/toolkit";
import { createSelector } from "reselect";
import { useDispatch, useSelector } from "react-redux";
import { setChosenProduct, setProducts, setRestaurant } from "./slice";
import { Product } from "../../../lib/types/products";
import { retrieveChosenProduct, retrieveProducts, retrieveRestaurant } from "./selector";
import { Member } from "../../../lib/types/member";
import ProductService from "../../services/ProductService";
import MemberService from "../../services/MemberService";
import { serverApi } from "../../../lib/config";

 const actiondispatch = (dispatch : Dispatch) => ({
  setRestaurant: (data: Member)=> dispatch(setRestaurant(data)),
    setChosenProduct: (data : Product) => dispatch(setChosenProduct(data)),
    
  });

  const chosenProductsRetriever = createSelector( retrieveChosenProduct, 
      (chosenProduct) => ({chosenProduct}) )

    const restaurantRetriever = createSelector( retrieveRestaurant, 
      (restaurant) => ({restaurant}) )


      interface ChosenProductsProps {
  onAdd: (input: any) => void;
}

export default function ChosenProduct(props: ChosenProductsProps) {
  const {onAdd} = props
  const { setChosenProduct, setRestaurant } = actiondispatch(useDispatch());
  
const { chosenProduct } = useSelector(chosenProductsRetriever);
const { restaurant } = useSelector(restaurantRetriever);

  const { productId } = useParams<{ productId: string }>();
  
  useEffect(() => {
    const product = new ProductService()
    product.getProduct(productId)
    .then((data) => setChosenProduct(data))
    .catch((err) => console.log("ChosenProduct:", err))

    const member = new MemberService()
    member.getRestaurant(productId)
    .then((data) => setRestaurant(data))
    .catch((err) => console.log("ChosenProduct:", err))
  }, [productId])
  if(!chosenProduct) return null
  return (

    <div className={"chosen-product"}>
      <Box className={"title"}>Product Detail</Box>
      <Container className={"product-container"}>
        <Stack className={"chosen-product-slider"}>
          <Swiper
            loop={true}
            spaceBetween={10}
            navigation={true}
            modules={[FreeMode, Navigation, Thumbs]}  
            className="swiper-area"
          >
            {chosenProduct.productImages.map(
              (ele: string, index: number) => {
                const imagePath = `${serverApi}/${ele.replace(/\\/g, "/")}`;
                return (
                  <SwiperSlide key={index}>
                    <img className="slider-image" src={imagePath} />
                  </SwiperSlide>
                );
              },
            )}
          </Swiper>
        </Stack>
        <Stack className={"chosen-product-info"}>
          <Box className={"info-box"}>
            <strong className={"product-name"}>{chosenProduct.productName}</strong>
            <span className={"resto-name"}>{restaurant?.memberNick}</span>
            <span className={"resto-name"}>{restaurant?.memberPhone}</span>
            <Box className={"rating-box"}>
              <Rating name="half-rating" defaultValue={2.5} precision={0.5} />
              <div className={"evaluation-box"}>
                <div className={"product-view"}>
                  <RemoveRedEyeIcon sx={{ mr: "10px" }} />
                  <span>{chosenProduct?.productViews || 0}</span>
                </div>
              </div>
            </Box>
            <p className={"product-desc"}>{chosenProduct?.productDesc || "No Description"}</p>
            <Divider height="1" width="100%" bg="#000000" />
            <div className={"product-price"}>
              <span>Price:</span>
              <span>${chosenProduct?.productPrice}</span>
            </div>
            <div className={"button-box"}>
              <Button variant="contained"
              onClick={(e) => {
                              onAdd({
                                _id: chosenProduct._id,
                                quantity: 1,
                                name: chosenProduct.productName,
                                image: chosenProduct.productImages[0],
                                price: chosenProduct.productPrice
                              })
                              e.stopPropagation()
                            }}>Add To Basket</Button>
            </div>
          </Box>
        </Stack>
      </Container>
    </div>
  );
}
